schnorr running in sp1
