+++ Cargo.toml
@@ -25,6 +25,7 @@ strata-native-zkvm-adapter = { path = "crates/native" }
 
 fibonacci = { path = "examples/fibonacci" }
 sha2-chain = { path = "examples/sha2-chain" }
+schnorr = { path = "examples/schnorr" }
 
 arbitrary = { version = "1.3.2", features = ["derive"] }
 bincode = "1.3"
