+++ artifacts/sp1/build.rs
@@ -3,5 +3,5 @@ use sp1_build::build_program;
 fn main() {
     build_program("fibonacci");
     build_program("sha2-chain");
-    build_program("schnorr-verification");
+    build_program("schnorr");
 }
