+++ artifacts/sp1/schnorr/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "guest-schnorr-verification"
+name = "guest-schnorr"
 version = "0.1.0"
 edition = "2021"
 
@@ -15,3 +15,4 @@ mock = ["strata-sp1-adapter/mock"]
 
 [patch.crates-io]
 sha2-v0-10-8 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", branch = "patch-sha2-v0.10.8" }
+secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
