+++ artifacts/sp1/schnorr/src/main.rs
