+++ runner/src/programs/mod.rs
@@ -1,6 +1,7 @@
 use std::str::FromStr;
 
 pub mod fibonacci;
+pub mod schnorr;
 pub mod sha2;
 
 #[derive(Debug)]
@@ -8,6 +9,7 @@ pub mod sha2;
 pub enum TestProgram {
     Fibonacci,
     Sha2Chain,
+    Schnorr,
 }
 
 impl FromStr for TestProgram {
@@ -17,6 +19,7 @@ impl FromStr for TestProgram {
         match s.to_lowercase().as_str() {
             "fibonacci" => Ok(TestProgram::Fibonacci),
             "sha2-chain" => Ok(TestProgram::Sha2Chain),
+            "schnorr" => Ok(TestProgram::Schnorr),
             // Add more matches
             _ => Err(format!("unknown program: {}", s)),
         }
